Move delete signal disabler to common utils
